directbuy solution
Answer Explanation :
First I have done with ForEach and  loop the array to find everything with parent id equal to id(compare array once for every element)
But for the 10,000 Categories, this is going to be slow because of the big O(n^2) where n will be 10,000 categories.
So the optimized solution is creating a lookup table and using key-value compare, this is going to be fast and big O(n) where n will be 10,000 categories. (solution1.js)
I have another solution using filter and findIndex and big O(n) where n will be 10,0000 categories. ( solution2.js)
I will send you the results in javascript code and 1 screenshot of test of the result in jsFiddle and paste it in the mail the 2 solutions.
I will in waiting for the evaluation of my answer.